pyproject.toml
[tool.poetry]
name = "passman"
version = "0.1.17"
description = "Cli tool to manage secrets on encrypted filesystem like kbfs"
authors = ["Bryan Neyedli <bryan@bledsol.net>"]
license = "BSD"
readme = "README.md"
packages = [
    { include = "passman", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
parseargs = "^0.0.5"
python-gnupg = "^0.5.0"
pathlib = "^1.0.1"
pyperclip = "^1.8.2"
pyotp = "^2.7.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
bandit = "^1.7.4"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.982"
pylint = "^2.15.4"
pytype = "^2022.10.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pman = 'passman.console.app:main'
